backend/api/requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
psycopg2-binary==2.9.5